feat: 코드리뷰 반영 및 MinIO 관련 파일들을 common폴더로 이동
