+++ src/main/java/until/the/eternity/dcs/common/exception/FileStorageException.java
@@ -1,0 +1,7 @@
+package until.the.eternity.dcs.common.exception;
+
+public class FileStorageException extends CustomException {
+    public FileStorageException(ExceptionCode code) {
+        super(code);
+    }
+}
