+++ src/main/java/until/the/eternity/dcs/common/exception/MissingFileUploadException.java
@@ -1,0 +1,9 @@
+package until.the.eternity.dcs.common.exception;
+
+import static until.the.eternity.dcs.common.exception.PostImageExceptionCode.MISSING_FILE_UPLOAD_EXCEPTION;
+
+public class MissingFileUploadException extends CustomException {
+    public MissingFileUploadException() {
+        super(MISSING_FILE_UPLOAD_EXCEPTION);
+    }
+}
